History screen refactor
